// src/entities/onboarding/model/constants.ts
export enum STEP_TYPES {
  Initial = 'initial',
  CheckImages = 'check_images',
  CheckButtonsColumn = 'check_buttons_column',
  CheckButtonsGrid = 'check_buttons_grid',
  BetweenResult = 'between_result',
  DateInput = 'date_input',
  TimeInput = 'time_input',
  TextInput = 'text_input',
  StepsProgress = 'steps_progress',
  Final = 'final',
}